Submit solution
So my approach for this was to store all possible roman number values in a hash map for simple roman to integer evaluation. Use two pointers(one for the previous position another for the current) to traverse the roman string, while traversing the array and add to the total sum while keeping track if we have found a pair. Keep track of pairs was required because if a pair was found it meant that the previous pointer was already added to the total sum, so a new previous pointer needed to be acquired. 

I did not test this algorithm extensively I only used the examples given within the readme